refactor network module
