+++ app/src/main/java/com/intive/selftraining/selftraining/di/module/NetworkModule.kt
@@ -1,7 +1,6 @@
 package com.intive.selftraining.selftraining.di.module
 
 import com.example.aleksandrtumanov.daggerattempts3.di.scopes.ActivityScope
-import com.intive.selftraining.selftraining.listmovies.ListMoviesRepository
 import com.intive.selftraining.selftraining.network.NetworkClient
 import com.intive.selftraining.selftraining.network.NetworkInterface
 import dagger.Module
@@ -13,11 +12,5 @@ class NetworkModule {
     @Provides
     fun provideApiService(): NetworkInterface {
         return NetworkClient().networkResponse
-    }
-
-    @ActivityScope
-    @Provides
-    fun provideRepo(networkInterface: NetworkInterface): ListMoviesRepository {
-        return ListMoviesRepository(networkInterface)
     }
 }
